requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-dotenv
pytest
pydantic
requests
python-multipart
sqlalchemy
psycopg2
